requirements.txt
langchain
langchain-google-genai
langchain-community
supabase
fastapi
uvicorn
streamlit
python-dotenv
pydantic
asyncpg
python-dateutil
sentence-transformers